feat: add vote object

+++ migrations/2022-12-09-021313_create_votes/up.sql
@@ -1,7 +1,7 @@
 -- Your SQL goes here
 CREATE TABLE votes (
     id uuid DEFAULT uuid_generate_v4(),
-    value SMALLINT NOT NULL UNIQUE,
+    value INTEGER NOT NULL UNIQUE,
     created_at TIMESTAMP NOT NULL DEFAULT NOW(),
     updated_at TIMESTAMP NOT NULL DEFAULT NOW(),
     user_id uuid NOT NULL,
